adorable_backend/requirements.txt
# Django and DRF
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1
django-redis==5.4.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Environment variables
python-dotenv==1.0.0

# Background Tasks
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1
hiredis==2.3.2

# Firebase Admin
firebase-admin==6.3.0

# Image Processing
Pillow==10.1.0

# Location Services
geopy==2.4.1
requests==2.31.0

# API Documentation
drf-yasg==1.21.7

# Monitoring and Logging
python-json-logger==2.0.7
psutil==5.9.7
sentry-sdk==1.39.1

# Security
cryptography==41.0.7
django-ratelimit==4.1.0

# Development Tools
black==23.12.1
flake8==7.0.0
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
mypy==1.8.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
uvicorn==0.25.0 